Fix innacurate test for whether or not a route was an IndexRoute
The issue was caused by the fact that we can't use reference equality to compare
an object to its minified self. Since we obviously want to minify in production
it is not viable to test `route === IndexRoute`.

The new approach is not very sophisticated, but if it causes issues down the
road we can address them.