Add support for minified bundles
This is an odd one. For some reason vm.runInThisContext was failing for the
minified bundle. As such, we needed a way to pass the unminifed source to the
vm. I ended up going with the asset._originalSource property since it provides
what we want, but I don't know if this is specific to the uglify plugin or if
it will present any issues down the road.

For instance, will it matter to react that the static source was rendered from
unminified code, while the bundle that gets sent to the browser is minified.

These are questions for later. For now, this works.