Remove hasNoComponent check
This check was causing an issue where an undefined component on a route would
cause any child routes to also not be rendered. In the specific example of #12
the root component was undefined. This meant that no routes were rendered
because they were all child routes of the root component.

This seems like a bit of an edge case, but all tests pass after this change so
it will be shipped.